add: make /recordings be the base for experiments & allow url filter by activityId
