+++ frontend/src/components/modals/index.tsx
@@ -1,0 +1,1 @@
+export * from "./onboarding-modal";
