+++ frontend/src/pages/recordings.tsx
@@ -1,13 +1,73 @@
 import { GetServerSideProps, NextPage } from "next";
 import { getServerSession } from "next-auth";
+import React, { useState, useEffect } from "react";
 
 import { authOptions } from "./api/auth/[...nextauth]";
 
-const RecodingPage: NextPage = () => {
-  return <div>Good, you&apos;re authenticated if you were able to make it here</div>;
+import { Experiment } from "~/components/lab";
+import { DashboardLayout, Meta } from "~/components/layouts";
+import { experiments } from "~/config/data";
+import { OnboardingModal } from "~/components/modals";
+import { useRouter } from "next/router";
+
+const RecordingsPage: NextPage = () => {
+  const [activeExperiment, setActiveExperiment] = React.useState(experiments[0]);
+
+  const router = useRouter();
+  const { activityId } = router.query;
+
+  useEffect(() => {
+    if (activityId && typeof activityId === "string") {
+      const chosenExperiment = experiments.find((experiment) => experiment.id === parseInt(activityId, 10));
+      if (chosenExperiment) {
+        setActiveExperiment(chosenExperiment);
+      }
+    }
+  }, [activityId]);
+
+  const handleExperimentSelect = (e: React.ChangeEvent<HTMLSelectElement>) => {
+    const chosenExperiment = experiments.find((experiment) => experiment.name === e.target.value);
+
+    if (chosenExperiment) {
+      setActiveExperiment(chosenExperiment);
+    }
+  };
+
+  return (
+    <DashboardLayout>
+      <Meta
+        meta={{
+          title: "Brain Recordings | NeuroFusion",
+          description:
+            "The simplest way to record and analyze your brain activity. Choose from a variety of cognitive experiments to record your brain activity and see results.",
+          image: "https://usefusion.app/images/features/neurofusion_experiment.png",
+        }}
+      />
+      <h1 className="text-4xl">Recordings</h1>
+      <label htmlFor="activity" className="my-2 block text-sm font-medium text-gray-900 dark:text-white">
+        Select activity:
+        <select
+          id="activity"
+          className="block w-full rounded-lg border border-gray-300 bg-gray-50 p-2.5 text-sm text-gray-900 focus:border-indigo-500 focus:ring-indigo-500 dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:placeholder-gray-400 dark:focus:border-indigo-500 dark:focus:ring-indigo-500"
+          onChange={handleExperimentSelect}
+          value={activeExperiment.name}
+        >
+          {experiments.map((experiment) => {
+            return (
+              <option key={experiment.id} value={experiment.name}>
+                {experiment.name}
+              </option>
+            );
+          })}
+        </select>
+      </label>
+      <Experiment {...activeExperiment} />
+      <OnboardingModal />
+    </DashboardLayout>
+  );
 };
 
-export default RecodingPage;
+export default RecordingsPage;
 
 export const getServerSideProps: GetServerSideProps = async ({ req, res }) => {
   const session = await getServerSession(req, res, authOptions);
